Move utility modules to tasks folder
