+++ config.default.yaml
@@ -1,7 +1,7 @@
 button_actions:
-  "0": time_utils.fetch_time
+  "0": tasks.time_utils.fetch_time
   "1":
-    func: weather.fetch_weather
+    func: tasks.weather.fetch_weather
     kwargs:
       area_code: "270000"
       area_name: "大阪"
